refactor: Moved event handling system into the new module event-system and dissolved the core/events folder.
