+++ src/app/core/abstractions/rundown-event-parser.interface.ts
@@ -6,7 +6,7 @@ import {
     RundownResetEvent,
     PartSetAsNextEvent,
     PartTakenEvent, RundownAdLibPieceInsertedEvent, RundownDeletedEvent
-} from '../../models/rundown-event'
+} from '../models/rundown-event'
 
 @Injectable()
 export abstract class RundownEventParser {
