+++ src/app/core/core.module.ts
@@ -6,18 +6,15 @@ import { HttpClientModule } from '@angular/common/http'
 import { HttpBasicRundownService } from './services/http-basic-rundown.service'
 import { HttpRundownService } from './services/http-rundown.service'
 import { AdLibPieceService } from './services/ad-lib-piece.service'
-import { ConnectionStatusObserver } from './services/events/connection-status-observer.service'
-import { EventObserver } from './services/events/event-observer.interface'
-import { WebSocketEventObserver } from './services/events/websocket-event-observer.service'
-import { RundownEventObserver } from './services/events/rundown-event-observer.service'
+import { ConnectionStatusObserver } from './services/connection-status-observer.service'
+import { RundownEventObserver } from './services/rundown-event-observer.service'
 import { RundownStateService } from './services/rundown-state.service'
-import { RobustWebSocketFactory } from './services/events/robust-websocket.factory'
-import { RundownEventParser } from './services/events/rundown-event-parser.interface'
+import { RundownEventParser } from './abstractions/rundown-event-parser.interface'
 import { BasicRundownStateService } from './services/basic-rundown-state.service'
 import { BasicRundownService } from './abstractions/basic-rundown-service'
 import { EntityParser } from './services/entity-parser.interface'
 import { ZodEntityParser } from './services/zod-entity-parser.service'
-import { ZodRundownEventParser } from './services/events/zod-rundown-event-parser.service'
+import { ZodRundownEventParser } from './parsers/zod-rundown-event-parser.service'
 
 @NgModule({
   declarations: [],
@@ -28,11 +25,9 @@ import { ZodRundownEventParser } from './services/events/zod-rundown-event-parse
   providers: [
     HttpErrorService,
     HttpClientModule,
-    RobustWebSocketFactory,
     HttpBasicRundownService,
     HttpRundownService,
     AdLibPieceService,
-    { provide: EventObserver, useClass: WebSocketEventObserver },
     ConnectionStatusObserver,
     RundownEventObserver,
     { provide: BasicRundownService, useClass: HttpBasicRundownService },
