+++ src/app/core/parsers/zod-rundown-event-parser.service.ts
@@ -1,6 +1,6 @@
-import { RUNDOWN_ACTIVATED_EVENT_PARSER } from '../../parsers/events/rundown-activated-event.schema'
+import { RUNDOWN_ACTIVATED_EVENT_PARSER } from './events/rundown-activated-event.schema'
 import { Injectable } from '@angular/core'
-import { RUNDOWN_TAKEN_EVENT_PARSER } from '../../parsers/events/rundown-taken-event.schema'
+import { RUNDOWN_TAKEN_EVENT_PARSER } from './events/rundown-taken-event.schema'
 import {
     RundownInfinitePieceAddedEvent,
     RundownActivatedEvent,
@@ -8,13 +8,13 @@ import {
     RundownResetEvent,
     PartSetAsNextEvent,
     PartTakenEvent, RundownAdLibPieceInsertedEvent, RundownDeletedEvent
-} from '../../models/rundown-event'
-import { RUNDOWN_DEACTIVATED_EVENT_PARSER } from '../../parsers/events/rundown-deactivated-event.schema'
-import { RUNDOWN_RESET_EVENT_PARSER } from '../../parsers/events/rundown-reset-event.schema'
-import { RUNDOWN_SET_NEXT_EVENT_PARSER } from '../../parsers/events/rundown-set-next-event.schema'
-import { RUNDOWN_AD_LIB_PIECE_INSERTED_EVENT_PARSER } from '../../parsers/events/rundown-adlib-piece-inserted-event.schema'
-import { RUNDOWN_INFINITE_PIECE_ADDED_EVENT_PARSER } from '../../parsers/events/rundown-infinite-piece-added-event.schema'
-import { RUNDOWN_DELETED_EVENT_PARSER } from '../../parsers/events/rundown-deleted-event.schema'
+} from '../models/rundown-event'
+import { RUNDOWN_DEACTIVATED_EVENT_PARSER } from './events/rundown-deactivated-event.schema'
+import { RUNDOWN_RESET_EVENT_PARSER } from './events/rundown-reset-event.schema'
+import { RUNDOWN_SET_NEXT_EVENT_PARSER } from './events/rundown-set-next-event.schema'
+import { RUNDOWN_AD_LIB_PIECE_INSERTED_EVENT_PARSER } from './events/rundown-adlib-piece-inserted-event.schema'
+import { RUNDOWN_INFINITE_PIECE_ADDED_EVENT_PARSER } from './events/rundown-infinite-piece-added-event.schema'
+import { RUNDOWN_DELETED_EVENT_PARSER } from './events/rundown-deleted-event.schema'
 
 @Injectable()
 export class ZodRundownEventParser {
