+++ src/app/core/services/basic-rundown-state.service.ts
@@ -1,8 +1,8 @@
 import { Injectable, OnDestroy } from '@angular/core'
 import { BehaviorSubject, lastValueFrom, SubscriptionLike } from 'rxjs'
-import { RundownEventObserver } from './events/rundown-event-observer.service'
-import { Unsubscribe } from './events/event-observer.interface'
-import { ConnectionStatusObserver } from './events/connection-status-observer.service'
+import { RundownEventObserver } from './rundown-event-observer.service'
+import { Unsubscribe } from '../../event-system/abstractions/event-observer.interface'
+import { ConnectionStatusObserver } from './connection-status-observer.service'
 import { BasicRundown } from '../models/basic-rundown'
 import { BasicRundownService } from '../abstractions/basic-rundown-service'
 import { RundownEvent } from '../models/rundown-event'
