+++ src/app/core/services/connection-status-observer.service.ts
@@ -1,4 +1,4 @@
-import { EventConsumer, EventObserver, TypedEvent, Unsubscribe } from './event-observer.interface'
+import { EventObserver, Unsubscribe } from '../../event-system/abstractions/event-observer.interface'
 import { Injectable, OnDestroy } from '@angular/core'
 
 enum ConnectionEventType {
