+++ src/app/core/services/rundown-event-observer.service.spec.ts
@@ -1,8 +1,8 @@
 import { RundownEventObserver } from './rundown-event-observer.service'
 import { anyString, anything, instance, mock, verify, when } from '@typestrong/ts-mockito'
-import { EventObserver } from './event-observer.interface'
-import { RundownEventParser } from './rundown-event-parser.interface'
-import { RundownEventType } from '../../models/rundown-event-type'
+import { EventObserver } from '../../event-system/abstractions/event-observer.interface'
+import { RundownEventParser } from '../abstractions/rundown-event-parser.interface'
+import { RundownEventType } from '../models/rundown-event-type'
 
 describe(RundownEventObserver.name, () => {
     describe(RundownEventObserver.prototype.subscribeToRundownActivation.name, () => {
