+++ src/app/core/services/rundown-state.service.ts
@@ -10,10 +10,10 @@ import { Rundown } from '../models/rundown';
 import { HttpRundownService } from './http-rundown.service';
 import { Segment } from '../models/segment';
 import { Part } from '../models/part';
-import { RundownEventObserver } from './events/rundown-event-observer.service'
-import { Unsubscribe } from './events/event-observer.interface'
+import { RundownEventObserver } from './rundown-event-observer.service'
+import { Unsubscribe } from '../../event-system/abstractions/event-observer.interface'
 import { ManagedSubscription } from './managed-subscription.service'
-import { ConnectionStatusObserver } from './events/connection-status-observer.service'
+import { ConnectionStatusObserver } from './connection-status-observer.service'
 
 @Injectable()
 export class RundownStateService implements OnDestroy {
