+++ src/app/event-system/abstractions/event-observer.interface.ts
