+++ src/app/event-system/factories/robust-websocket.factory.ts
