+++ src/app/event-system/services/robust-websocket.service.ts
